refactor: cut  all Dao about Customer to dao directory
