+++ src/main/java/org/example/techcare/model/laptop/LaptopDAO.java
@@ -3,7 +3,7 @@ package org.example.techcare.model.laptop;
 import org.example.techcare.dao.BrandLaptopDAO;
 import org.example.techcare.model.brandlaptop.BrandLaptop;
 import org.example.techcare.model.customers.Customers;
-import org.example.techcare.model.customers.CustomersDAO;
+import org.example.techcare.dao.CustomersDAO;
 import org.example.techcare.model.utils.ConnectionBdd;
 
 import java.sql.PreparedStatement;
